Move modals into views folder
