+++ src/views/mod.rs
@@ -1,10 +1,13 @@
 pub mod configuration_column;
+#[cfg(any(feature = "iroh", feature = "tcp"))]
+pub mod connect_dialog_handler;
 pub mod hardware_menu;
 pub mod hardware_view;
 pub mod info_row;
 pub mod layout_selector;
 pub mod main_row;
 pub mod message_row;
+pub mod modal_handler;
 pub mod pin_state;
 pub mod unsaved_status;
 pub mod version;
